services/user-service/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.37.2
botocore==1.37.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.113.0
fastapi-cli==0.0.5
grpcio==1.66.2
grpcio-tools==1.66.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
jmespath==1.0.1
mangum==0.17.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
protobuf==5.27.2
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.0
s3transfer==0.11.3
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.38.4
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
